requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
openai>=1.0.0
Pillow>=10.0.0
colorgram.py>=1.2.0
PyYAML>=6.0
python-dotenv>=1.0.0
